+++ mycodo/inputs/mcp342x.py
@@ -83,7 +83,7 @@ class InputModule(AbstractInput):
     def get_measurement(self):
         self.return_dict = measurements_dict.copy()
 
-        for channel in self.device_measurements:
+        for channel in self.channels_measurement:
             if self.is_enabled(channel):
                 adc = self.MCP342x(self.bus,
                                    self.i2c_address,